Fix radix sort on Mac
Fixes a __local variable which must be declared at __kernel function
scope (only the Mac complains about this, but it seems to be that
way in the spec too). Has the nice effect of removing a Mac-specific
ifdef from the kernel code which wasn't being used correctly.
Also added some code to check allowed work group sizes.